requirements.txt
configparser==3.5.0
Django==1.9
django-extensions==1.7.5
djangorestframework==3.4.7
enum34==1.1.6
flake8==3.2.1
mccabe==0.5.3
pycodestyle==2.2.0
pyflakes==1.3.0
requests==2.12.5
six==1.10.0
wheel==0.24.0
